app/requirements.txt
tensorflow==2.18.0
streamlit==1.30.0
numpy>=1.26.0,<2.1.0  # ✅ Allows compatible versions
Pillow
gdown
